+++ collect_app/src/main/java/org/odk/collect/android/preferences/screens/AdminPreferencesFragment.kt
@@ -15,7 +15,6 @@ package org.odk.collect.android.preferences.screens
 
 import android.content.Intent
 import android.os.Bundle
-import androidx.fragment.app.Fragment
 import androidx.preference.Preference
 import org.odk.collect.android.R
 import org.odk.collect.android.utilities.MultiClickGuard
@@ -27,9 +26,6 @@ class AdminPreferencesFragment :
         super.onCreatePreferences(savedInstanceState, rootKey)
         setPreferencesFromResource(R.xml.admin_preferences, rootKey)
         findPreference<Preference>("odk_preferences")!!.onPreferenceClickListener = this
-        findPreference<Preference>("main_menu")!!.onPreferenceClickListener = this
-        findPreference<Preference>("user_settings")!!.onPreferenceClickListener = this
-        findPreference<Preference>("form_entry")!!.onPreferenceClickListener = this
     }
 
     override fun onPreferenceClick(preference: Preference): Boolean {
@@ -40,23 +36,10 @@ class AdminPreferencesFragment :
                     intent.putExtra(GeneralPreferencesActivity.INTENT_KEY_ADMIN_MODE, true)
                     startActivity(intent)
                 }
-                "main_menu" -> displayPreferences(MainMenuAccessPreferencesFragment())
-                "user_settings" -> displayPreferences(UserSettingsAccessPreferencesFragment())
-                "form_entry" -> displayPreferences(FormEntryAccessPreferencesFragment())
             }
             return true
         }
         return false
-    }
-
-    private fun displayPreferences(fragment: Fragment?) {
-        if (fragment != null) {
-            fragment.arguments = arguments
-            requireActivity().supportFragmentManager.beginTransaction()
-                .replace(R.id.preferences_fragment_container, fragment)
-                .addToBackStack(null)
-                .commit()
-        }
     }
 
     fun preventOtherWaysOfEditingForm() {
