+++ collect_app/src/main/res/xml/access_control_preferences.xml
@@ -1,0 +1,22 @@
+<?xml version="1.0" encoding="utf-8"?>
+<PreferenceScreen xmlns:android="http://schemas.android.com/apk/res/android"
+    android:title="@string/access_control_section_title">
+
+    <Preference
+        android:icon="@drawable/ic_outline_menu_24"
+        android:key="main_menu"
+        android:summary="@string/main_menu_settings_summary"
+        android:title="@string/main_menu_settings" />
+
+    <Preference
+        android:icon="@drawable/ic_outline_settings_24"
+        android:key="user_settings"
+        android:summary="@string/user_setting_summary"
+        android:title="@string/user_settings" />
+
+    <Preference
+        android:icon="@drawable/ic_outline_assignment_24"
+        android:key="form_entry"
+        android:summary="@string/form_entry_settings_summary"
+        android:title="@string/form_entry_setting" />
+</PreferenceScreen>
