+++ strings/src/main/res/values-ar/strings.xml
@@ -718,7 +718,6 @@
   <string name="reset_blank_forms_result">الاستمارات الفارغة  :: %s</string>
   <string name="reset_layers_result">طبقات الخرائط  :: %s</string>
   <string name="reset_cache_result">التخزين المؤقت للنماذج المحملة  :: %s</string>
-  <string name="user_access_control">قيود المستخدم</string>
   <string name="main_menu_settings">إعدادات القائمة الرئيسة</string>
   <string name="main_menu_settings_summary">إنقر لإدارة عناصر القائمة الرئيسية</string>
   <string name="user_settings">إعدادات المستخدم</string>
