+++ strings/src/main/res/values-cs/strings.xml
@@ -712,7 +712,6 @@
   <string name="reset_blank_forms_result">Prázdné formuláře: %s</string>
   <string name="reset_layers_result">Mapové vrstvy :: %s</string>
   <string name="reset_cache_result">Mezipaměť pro načtení formuláře :: %s</string>
-  <string name="user_access_control">Řízení přístupu uživatelů</string>
   <string name="main_menu_settings">Nastavení hlavního menu</string>
   <string name="main_menu_settings_summary">Klepnutím spravovat položky hlavní nabídky</string>
   <string name="user_settings">Uživatelské nastavení</string>
