+++ strings/src/main/res/values-hi/strings.xml
@@ -467,7 +467,6 @@
   <string name="reset_blank_forms_result">खाली  फॉर्म:: %s</string>
   <string name="reset_layers_result">मानचित्र लेयरें :: %s</string>
   <string name="reset_cache_result">फॉर्म लोड कैश :: %s</string>
-  <string name="user_access_control">प्रयोगकर्ता एक्सेस नियंत्रण</string>
   <string name="main_menu_settings">मुख्य मेनू सेटिंग्स</string>
   <string name="main_menu_settings_summary">मुख्य मेनू आइटम मैनेज करने के लिए टैप करें</string>
   <string name="user_settings">प्रयोगकर्ता सेटिंग</string>
