+++ strings/src/main/res/values-in/strings.xml
@@ -703,7 +703,6 @@
   <string name="reset_blank_forms_result">Formulir kosong :: %s</string>
   <string name="reset_layers_result">Lapisan peta :: %s</string>
   <string name="reset_cache_result">Cache formulir yang dimuat :: %s</string>
-  <string name="user_access_control">Kontrol Akses Pengguna</string>
   <string name="main_menu_settings">Pengaturan Menu Utama</string>
   <string name="main_menu_settings_summary">Sentuh untuk mengelola item menu utama</string>
   <string name="user_settings">Pengaturan Pengguna</string>
