+++ strings/src/main/res/values-pl/strings.xml
@@ -662,7 +662,6 @@
   <string name="reset_blank_forms_result">Puste formularze :: %s</string>
   <string name="reset_layers_result">Warstwy mapy :: %s</string>
   <string name="reset_cache_result">Pamięć podręczna załadowanego formularza :: %s</string>
-  <string name="user_access_control">Kontrola dostępu użytkownika</string>
   <string name="main_menu_settings">Ustawienia Menu Głównego</string>
   <string name="main_menu_settings_summary">Dotknij, by zarządzać elementami menu głównego</string>
   <string name="user_settings">Ustawienia użytkownika</string>
