+++ strings/src/main/res/values-sq/strings.xml
@@ -495,7 +495,6 @@
   <string name="reset_blank_forms_result">Formularë bosh :: %s</string>
   <string name="reset_layers_result">Shtresat e hartës :: %s</string>
   <string name="reset_cache_result">Cache ngarkimi i formularëve :: %s</string>
-  <string name="user_access_control">Kontrolli i aksesi të përdoruesit</string>
   <string name="main_menu_settings">Parametrat e Menusë Kryesore</string>
   <string name="main_menu_settings_summary">Administro elementët e menusë kryesore</string>
   <string name="user_settings">Parametrat e përdoruesit</string>
